[DOM-239] Change Auth module way of navigation
